+++ app/main-view-model.js
@@ -1,7 +1,7 @@
 var Observable = require("data/observable").Observable;
 var application = require("application");
 var utils = require("utils/utils");
-var services = require("./service-helper");
+var jobScheduler = require("./notifications/job-scheduler");
 
 function getMessage(counter) {
     return counter;
@@ -12,7 +12,7 @@ function createViewModel() {
     viewModel.message = "Schedule a local notification every 20 mins";
 
     viewModel.onTap = function () {
-        services.scheduleJob(utils.ad.getApplicationContext());
+        jobScheduler.scheduleJob(utils.ad.getApplicationContext());
     }
 
     return viewModel;
